Implement TakeCount, bump to version 2.0

+++ examples/discard_lines.rs
@@ -6,8 +6,8 @@ use token_read::TokenReader;
 fn main() -> Result<()> {
     let mut input = TokenReader::new(stdin().lock());
 
-    let (line_count, _discard_count): (usize, usize) = input.line()?;
-    let discards: HashSet<usize> = input.line()?;
+    let (line_count, _discard_count): (u64, usize) = input.line()?;
+    let discards: HashSet<u64> = input.line()?;
 
     for i in 0..line_count {
         let line = input.line_raw()?;
